+++ src/services/operator-date-interval-parser.js
@@ -46,7 +46,7 @@ const DATE_OPERATORS = [
 ];
 
 function OperatorDateIntervalParser(timezone, options) {
-  const offsetClient = parseInt(moment().tz(timezone).format('Z'), 10);
+  const offsetClient = Number.parseInt(moment().tz(timezone).format('Z'), 10);
   const offsetServer = moment().utcOffset() / 60;
 
   this.offsetHours = offsetServer - offsetClient;
@@ -60,105 +60,122 @@ function OperatorDateIntervalParser(timezone, options) {
     .includes(operator);
 
   this.getDateIntervalFilter = (operator, value) => {
-    const condition = {};
+    value = Number.parseInt(value, 10);
 
     switch (operator) {
       case 'today':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment().startOf('day'));
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment().endOf('day'));
-        break;
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(moment().startOf('day')),
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(moment().endOf('day')),
+        };
       case 'past':
-        condition[this.OPERATORS.LTE] = moment().toDate();
-        break;
+        return { [this.OPERATORS.LTE]: moment().toDate() };
       case 'future':
-        condition[this.OPERATORS.GTE] = moment().toDate();
-        break;
+        return { [this.OPERATORS.GTE]: moment().toDate() };
       case 'yesterday':
       case 'previous_week':
       case 'previous_month':
       case 'previous_quarter':
-      case 'previous_year':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .subtract(1, PERIODS[operator]).startOf(PERIODS_VALUES[PERIODS[operator]]));
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment()
-          .subtract(1, PERIODS[operator]).endOf(PERIODS_VALUES[PERIODS[operator]]));
-        break;
+      case 'previous_year': {
+        const previousPeriod = moment().subtract(1, PERIODS[operator]);
+
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(previousPeriod
+            .startOf(PERIODS_VALUES[PERIODS[operator]])),
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(previousPeriod
+            .endOf(PERIODS_VALUES[PERIODS[operator]])),
+        };
+      }
       case 'previous_week_to_date':
       case 'previous_month_to_date':
       case 'previous_quarter_to_date':
       case 'previous_year_to_date':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .startOf(PERIODS_VALUES[PERIODS[operator]]));
-        condition[this.OPERATORS.LTE] = moment().toDate();
-        break;
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(moment()
+            .startOf(PERIODS_VALUES[PERIODS[operator]])),
+          [this.OPERATORS.LTE]: moment().toDate(),
+        };
       case 'previous_x_days':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .subtract(value, 'days').startOf('day'));
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment()
-          .subtract(1, 'days').endOf('day'));
-        break;
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(moment()
+            .subtract(value, 'days').startOf('day')),
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(moment()
+            .subtract(1, 'days').endOf('day')),
+        };
       case 'previous_x_days_to_date':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .subtract(value - 1, 'days').startOf('day'));
-        condition[this.OPERATORS.LTE] = moment().toDate();
-        break;
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(moment()
+            .subtract(value - 1, 'days').startOf('day')),
+          [this.OPERATORS.LTE]: moment().toDate(),
+        };
       case 'before_x_hours_ago':
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment().subtract(value, 'hours'));
-        break;
+        return {
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(moment().subtract(value, 'hours')),
+        };
       case 'after_x_hours_ago':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment().subtract(value, 'hours'));
-        break;
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(moment().subtract(value, 'hours')),
+        };
       default:
         throw new NoMatchingOperatorError();
     }
-    return condition;
   };
 
   this.getPreviousDateIntervalFilter = (operator, value) => {
-    const condition = {};
+    value = Number.parseInt(value, 10);
 
     switch (operator) {
-      case 'today':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .subtract(1, 'days').startOf('day'));
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment()
-          .subtract(1, 'days').endOf('day'));
-        break;
+      case 'today': {
+        const yesterday = moment().subtract(1, 'days');
+
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(yesterday.startOf('day')),
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(yesterday.endOf('day')),
+        };
+      }
       case 'previous_x_days':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .subtract(value * 2, 'days').startOf('day'));
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment()
-          .subtract(value + 1, 'days').endOf('day'));
-        break;
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(moment()
+            .subtract(value * 2, 'days').startOf('day')),
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(moment()
+            .subtract(value + 1, 'days').endOf('day')),
+        };
       case 'previous_x_days_to_date':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .subtract((value * 2) - 1, 'days').startOf('day'));
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment()
-          .subtract(value, 'days'));
-        break;
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(moment()
+            .subtract((value * 2) - 1, 'days').startOf('day')),
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(moment()
+            .subtract(value, 'days')),
+        };
       case 'yesterday':
       case 'previous_week':
       case 'previous_month':
       case 'previous_quarter':
-      case 'previous_year':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .subtract(2, PERIODS[operator]).startOf(PERIODS_VALUES[PERIODS[operator]]));
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment()
-          .subtract(2, PERIODS[operator]).endOf(PERIODS_VALUES[PERIODS[operator]]));
-        break;
+      case 'previous_year': {
+        const penultimatePeriod = moment().subtract(2, PERIODS[operator]);
+
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(penultimatePeriod
+            .startOf(PERIODS_VALUES[PERIODS[operator]])),
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(penultimatePeriod
+            .endOf(PERIODS_VALUES[PERIODS[operator]])),
+        };
+      }
       case 'previous_week_to_date':
       case 'previous_month_to_date':
       case 'previous_quarter_to_date':
-      case 'previous_year_to_date':
-        condition[this.OPERATORS.GTE] = this.toDateWithTimezone(moment()
-          .subtract(1, PERIODS[operator]).startOf(PERIODS_VALUES[PERIODS[operator]]));
-        condition[this.OPERATORS.LTE] = this.toDateWithTimezone(moment()
-          .subtract(1, PERIODS[operator]));
-        break;
+      case 'previous_year_to_date': {
+        const previousPeriod = moment().subtract(1, PERIODS[operator]);
+
+        return {
+          [this.OPERATORS.GTE]: this.toDateWithTimezone(previousPeriod
+            .startOf(PERIODS_VALUES[PERIODS[operator]])),
+          [this.OPERATORS.LTE]: this.toDateWithTimezone(previousPeriod),
+        };
+      }
       default:
         throw new NoMatchingOperatorError();
     }
-    return condition;
   };
 }